fix bug: dashboard produk user pindah ke component
